requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
idna==2.10
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
plotly==5.0.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
smith-the-crawler==0.0.10a0
soupsieve==2.2.1
tenacity==7.0.0
urllib3==1.26.5
